Introduce user tokens for handling auth
This commit pulls some of the changes brought in Phoenix 1.8 [1]
around authentication. Notably, when a user logs in, we create a user
token that we store in the DB and associate with the session cookie.
This enables us to do two things:

  a) introduce a "sudo mode" where certain pages (such as account deletion)
can only be accessed if the user token has been generated recently and
present the login screen if not.

  b) revoke server side old sessions (for example when a user changes
  their password).

[1] https://hexdocs.pm/phoenix/1.8.0-rc.0/mix_phx_gen_auth.html#migrating-to-phoenix-v1-8-magic-links-and-sudo-mode