Relax query for public_links materialized view
This corrects a bug that was introduced recently with the total number
of shares being incorrectly shown. The issue was that the materialized
view containing public links was too strict in its selection query,
omitting links created by unconfirmed users and/or through IFTTT
integration.

The public_links materialized view should include all publicly
accessible links.
To prevent bookmarks from unconfirmed accounts and/or inserted via ifttt
in the homepages (recent and popular), the querying of the view has to
be updated, but not the underlying materialized view.