Infer title from the URL when adding bookmark
When a bookmark is added without a title (left blank in submission form)
then linkhut will attempt to determine the title from the provided URL.

This roughly works as follows:

If the scheme of the URL is either https:// or http://, then linkhut
requests the page. If the content-type of the response is "text/html"
then we look for a title by looking for the following tags (in this
order):
    1. the open graph title tag (og:title)
    2. the document title element (`<title>` tag)
    3. the first level-1 section heading (`<h1>` tag)

In the case that no title can be found using this approach, then the
path of the URL is used as a title, and if there's none then the domain.

Fixes: https://todo.sr.ht/~mlb/linkhut/34